# services/server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leasy Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quiz/questions/{uuid}": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Get the questions of the quiz",
                "description": "Fetches all questions of a quiz.",
                "operationId": "0365c35651f8003e815d93e90099a89a",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the quiz",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Questions retrieved successfully"
                    },
                    "404": {
                        "description": "Quiz not found"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/quiz/answer/{uuid}": {
            "put": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Answer quiz",
                "description": "Submits answers for a quiz",
                "operationId": "38686818dfd03ee9489dae4c0bd15c43",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the quiz",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Answers for the quiz questions",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answers": {
                                        "description": "List of answers",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "question_uuid",
                                                "answer"
                                            ],
                                            "properties": {
                                                "question_uuid": {
                                                    "description": "The UUID of the question",
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "answer": {
                                                    "type": "string",
                                                    "example": "2"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answers submitted successfully"
                    },
                    "400": {
                        "description": "This question cannot be answered"
                    },
                    "404": {
                        "description": "Some of the quiz properties not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
...
        "/api/video/last-watched-time/{uuid}": {
            "put": {
                "tags": [
                    "Videos"
                ],
                "summary": "Update the last watched time for a video",
                "description": "Updates the last watched time for a specific video using the provided video UUID.",
                "operationId": "334262d4225b3be95a8471048d7378cb",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the video to update the last watched time",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The last watched time to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "last_watched_time": {
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Last watched time updated successfully"
                    },
                    "404": {
                        "description": "Video not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/video/fix-audio/{uuid}": {
            "put": {
                "tags": [
                    "Videos"
                ],
                "summary": "Fix audio of a video",
                "description": "Fixes the audio issues of the specified video.",
                "operationId": "90e7b09d4d3df597c76fa28772796e26",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the video to fix audio",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio fixed successfully"
                    },
                    "404": {
                        "description": "Video not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    },
                    "204": {
                        "description": "No content"
                    }
                }
            }
        }
    }
}